libril: Work around idiotic operator names
libril-qc has a very messed up table of operator names, with the companies
formal names (which in many cases aren't recognizable to the end-user) as
the PLMN instead of their market brands. The SPNs are OK, though, and this
patch tries to force the display of that value whenever possible.